Renamed namespace for convenience
Renamed CurrencyGraph namespace to Graph for convenience.
