+++ Common/QuantConnect.csproj
@@ -176,10 +176,10 @@
     <Compile Include="Data\Custom\Intrinio\IntrinioEconomicData.cs" />
     <Compile Include="Data\Custom\Intrinio\EconomicDataSources.cs" />
     <Compile Include="Data\Fundamental\Fundamentals.cs" />
-    <Compile Include="Data\Graph\CurrencyPath.cs" />
-    <Compile Include="Data\Graph\StringEdge.cs" />
-    <Compile Include="Data\Graph\CurrencyGraph.cs" />
-    <Compile Include="Data\Graph\StringVertex.cs" />
+    <Compile Include="Securities\Graph\CurrencyPath.cs" />
+    <Compile Include="Securities\Graph\StringEdge.cs" />
+    <Compile Include="Securities\Graph\CurrencyGraph.cs" />
+    <Compile Include="Securities\Graph\StringVertex.cs" />
     <Compile Include="Data\UniverseSelection\ITimeTriggeredUniverse.cs" />
     <Compile Include="Data\UniverseSelection\ScheduledUniverse.cs" />
     <Compile Include="Exceptions\DllNotFoundPythonExceptionInterpreter.cs" />
