+++ Common/Securities/Graph/StringVertex.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace QuantConnect.Securities.CurrencyGraph
+namespace QuantConnect.Securities.Graph
 {
     public class CurrencyVertex
     {
